sug_crawler/requirements.txt
asyncpg
gunicorn
python-dotenv
rabbitpy
flask
flask-sqlalchemy
aiosqlite
aiohttp
aiosocksy
psycopg2